Added elastic filters

+++ store/elasticsearch.js
@@ -5,9 +5,7 @@ BE CAREFUL! Work in progress! This still unstable and unfinished!
 
 TO DO:
 
- - Add get function
  - Document all the functions (jsdocs)
- - Study insert vs update use case
  - Create the search model converter
  - Standarize callbacks (see datastore.js)
 
@@ -20,44 +18,93 @@ const config = require('../config');
 const logger = require('../logger');
 
 const INDEX = config.elastic_index;
-var login = config.elastic_user ? (config.elastic_pass ?  config.elastic_user + ':' +  config.elastic_pass + '@') : config.elastic_user + '@' : '';
+var login = config.elastic_user ? ( config.elastic_pass ? config.elastic_user + ':' + config.elastic_pass + '@' : config.elastic_user + '@' ) : '';
 
 // connect elasticsearch
 var client = new elasticsearch.Client({
-	host: 'http://' + login + config.elastic_host + ':' + elastic_port,
-	// log: config.elastic_log,
+	host: 'http://' + login + config.elastic_host + ':' + config.elastic_port,
+	log: config.elastic_log,
 	// log: 'error',
-	log: ['error', 'debug', 'info'],
 	maxRetries: 7,
 	maxSockets: 50,
 	keepAlive: true
 });
 
 /**
- *	 
+ * Callback for elastic get function.
+ * 
+ * @callback elasticGetCallback
+ * @param {object} data Data gotten or null
+ */
+/**
+ * Get a document by type and ID
+ *
+ * @param {string} type	Document type (sql "table")
+ * @param {string} id	Document ID (sql "primary key")
+ * @param {elasticGetCallback} cb 	Callback on get (or error)
  */
 function get(type, id, cb) {
+	if(!type || !id) {
+		if(cb && typeof cb) {
+			cb();
+		}
+		return false;
+	}
 
-	// Look for an example of this query
+	client.get({
+		index: INDEX,
+		type: type,
+		id: id
+	}, function (error, response) {
+		if(error) {
+			logger.error(error);
+		}
 
-	cb();
+		if(response && response.found && response._source){
+			cb(response._source);
+		} else {
+			cb(null);
+		}
+	});
 }
 
 /**
- *	 
+ * Callback for elastic upsert function.
+ * 
+ * @callback elasticUpsertCallback
+ * @param {boolean} success True if upserted, false on error
+ * @param {string} id Id of the upserted element
+ */
+/**
+ * Insert or update data in elasticseatch
+ *
+ * @param {string} type	Document type (sql "table")
+ * @param {object} data Data to write
+ * @param {string} id	Document ID (sql "primary key")
+ * @param {elasticUpsertCallback} cb Callback on upsert (or error)
  */
 function upsert(type, data, id, cb) {
+	if(!type || !data) {
+		if(cb && typeof cb) {
+			cb();
+		}
+		return false;
+	}
 
-	// TODO:
-	// Check if index exists or not
-	
-	// Add an index
-	client.index({
+	const options = {
 		index: INDEX,
 		type: type,
-		id: id,
 		body: data
-	}, function (error, response) {
+	};
+
+	if(id && typeof id !== 'function') {
+		options.id = id;
+	} else if(typeof id === 'function') {
+		cb = id;
+	}
+
+	// Add an index
+	client.index(options, function (error, response) {
 		if(error) {
 			logger.error('Have been an error writting in ElasticSearch!!');
 			logger.error(error);
@@ -67,14 +114,33 @@ function upsert(type, data, id, cb) {
 		}
 
 		// logger.log(response);
-		cb(response);
+		if(cb && typeof cb) {
+			cb(response);
+		}
 	});
 }
 
 /**
- *	 
+ * Callback for elastic remove function.
+ * 
+ * @callback elasticRemoveCallback
+ * @param {boolean} success True if removed, false on error
+ */
+/**
+ * Remove data in elasticseatch by ID
+ *
+ * @param {string} type	Document type (sql "table")
+ * @param {string} id	Document ID (sql "primary key")
+ * @param {elasticUpsertCallback} cb Callback on remove (or error)
  */
 function remove(type, id, cb) {
+	if(!type || !id) {
+		if(cb && typeof cb) {
+			cb();
+		}
+		return false;
+	}
+
 	client.delete({
 		index: INDEX,
 		type: type,
@@ -93,25 +159,71 @@ function remove(type, id, cb) {
 }
 
 /**
- *	 
+ * Callback for elastic query function.
+ * 
+ * @callback elasticRemoveCallback
+ * @param {array} results List of entities matching query
+ */
+/**
+ * Query data in elasticseatch by ID
+ *
+ * @param {string} type	Document type (sql "table")
+ * @param {object} options	Query options
+ * @param {elasticUpsertCallback} cb Callback on remove (or error)
  */
 function query(type, options, cb) {
+	if(!type || !options) {
+		if(cb && typeof cb) {
+			cb();
+		}
+		return false;
+	}
+
+	var body = options;
+
+	if(typeof options === 'function' && !cb) {
+		cb = options;
+		body = {}
+	} 
+	
 
 	// TODO:
 	// Convert our query model to elastic model
+	if (options.filters) {
+		body.query = {bool: { must: [] }}; 
+
+		for(let filter in option filters) {
+			console.log(filters[filter]);
+		}
+	}
+
+	body = {
+		query: {
+			bool: {
+				must: [
+					{ match: { type: 'create' } }
+				]
+			}
+		},
+		sort: {
+			added: { order: "desc" }
+		},
+		size: 3
+	}
+
+	const filters = options.filters || null;
+	const groupBy = options.groupBy || null;
+	const limit = options.limit || null;
+	const offset = options.offset || null;
+	const orderBy = options.orderBy || null;
+
+	// --------------- 
+
 
 	client.search({
 		index: INDEX,
 		type: type,
-		body: {
-			query: {
-				match: { type: 'create' } 
-			},
-			sort: {
-				added: { order: "desc" }
-			},
-			size: 3
-		}
+		body: body
 	}, function(error, response) {
 		
 		// Log errors
@@ -132,13 +244,68 @@ function query(type, options, cb) {
 			searchArray.push(response.hits.hits[i]._source);
 		}
 
-		cb(searchArray);
+		if(cb && typeof cb) {
+			cb(searchArray);
+		}
 
 	});
-
-	// cb();
 }
 
+/* [internal] parse filters to elasticsearch query syntax */
+function filters(body, options) {
+	if (options.filters) {
+		body.query = { bool: {} };
+		for(var filter in options.filters) {
+			var f = options.filters[filter];
+			var e = {};
+
+			switch (f.operator) {
+				case '=':
+				case '!=':
+					e.match = {};
+					e.match[f.field] = f.value
+					break;
+
+				case '>':
+				case '>=':
+				case '<=':
+				case '<':
+					var op = '';
+					if (f.operator === '>') {
+							op = 'gt';
+					} else if (f.operator === '>=') {
+							op = 'gte';
+					} else if (f.operator === '<') {
+							op = 'lt';
+					} else if (f.operator === '<=') {
+							op = 'lte';
+					}
+
+					e.range = {};
+					e.range[f.field] = {}
+					e.range[f.field][op] = f.value;
+					break;
+
+				default:
+					continue;
+
+			}
+
+			if(f.operator === '!=') {
+				if (typeof body.query.bool.must_not === 'undefined') {
+					body.query.bool.must_not = [];
+				}
+				body.query.bool.must_not.push(e);
+			} else {
+				if (typeof body.query.bool.must === 'undefined') {
+					body.query.bool.must = [];
+				}
+				body.query.bool.must.push(e);
+			}
+
+		}
+	}
+}
 
 module.exports = {
 	get: get,
